backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
gunicorn==23.0.0
python-dotenv==1.0.1
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
python-jose==3.3.0
pydantic==2.10.4
pydantic-settings==2.7.1
requests==2.32.3
